Don't trust the grid, compute the average density (#6897)
* Don't trust the grid, compute the average density

WHen grdflexure is used with a variable density grid (-Hrho.grd) we try to read a specific attribute in the remakr that holds the average density of the grid.  However, there is no knowing that this will be set properly or at all.  Hence, this PR always computes the mean density and, if present, compares with the attribute value and gives a warning if they differ by more than some small amount.

* Update grdflexure.c

* Root density option and handling