Tolerate round-off in z-array (#6903)
* Tolerate round-off in z-array

When grdinterpolate merges grids into a cube given a z-array file (via -Z), we checked if the array was equidstant or not.  That test was too harsh (used doubleAlmostEqual) but because of natural roundoff we need to allow slope slop.  This PR computes the successive dz_i/dz_i+1 remainders and if that exceeds 10e-12 then we fail, else it is equidistant.

* Update grdinterpolate.c

* Update grdinterpolate.c