Fix autoscaling -JX for double negatives
When -JX was given one negative length and another as -0 then the axis with -0 scale did not properly set which direction the axis was going, leading to bad translations and annotations on the wrong side.  This PR fixes this and adds a test that checks all possible -JX combinations involving ±0 as one axis width.