Incorrect comparison in gmt_grd_domains_match function
This function is sued to check if two grids have the same domain and increment and registration.  It takes a comment string to give the error message context. Unfortunately, there check for NULL was instead an assignment, leading to messages like

gravfft [ERROR]: The (null) grids have different regions

This PR fixes this so we now get

gravfft [ERROR]: The surface and density grids have different regions