Let grdimage handle mask grids
If you ceate a mask grid (NaN and 1) and want to see how it looks iin grdimage you get this annoyuance:

(base) pwessel@MacAttack-> gmt grdimage E_obl_inside_mask.grd -B -pdf t
grdimage [ERROR]: Passing zmax <= zmin prevents automatic CPT generation!
grdimage [ERROR]: Failed to read CPT (null).
psconvert [ERROR]: Unable to decode BoundingBox file /Users/pwessel/.gmt/sessions/gmt_session.11534/psconvert_30470c.bb (maybe no non-white features were plotted?)

and a failed plot.   The reason is that xmin = zmax = 1 and no CPT can be crafted.  Since this is a common situation, this PR allwos for this by checking if

We have a grid with NaNs
The rest of the data are all 1.0

If so it sets the internal zmax to 2 so that we get a map when no CPT is given.