Allow batch -F to set a more specific naming template (#6934)
* Let batch -D drop the mandatory moving of product files

See #6932 for discussion.  This PR implements a -D option that stops any copying.  This solution does not preempt further consideration of the other suggestion which is to generalize the naming of product files via a -Ftemplate option.

* Update batch.c

* Allow batch _F to set a more specific naming template

Rather than have a single running number as the product prefix we build one from this template and the -T data.

* Add the -Ftemplate option

* Update batch.rst

* polish

* Update batch.c

* Replace filename spaces in BATCH_NAME w underscore

* Update batch.rst

* Update batch.c