Quiet an inaccurate complaint about subshell (#6957)
* Add required check in gmtinit_get_region_from_data

See #6955 for background. The reason for the crash is that the malformed input record sets n col = 1 and then later when we run gmt info on it to detect -R we failed to check properly that we got 4 data columns back (xmin, xmax, ymin, ymax).  In fact, we pass the hidden -) option to also return the data column types and our check failed to detect this problem.

* Quiet an inaccurate complaint about subshell

When modern bash syntax using (( ...)) is used the $ identifying a variable is not needed, hence the affected test should check for ((.  No effect except on that message.