Fix mistake in using +1|2 instead of +d1|2 in grdinterpolate -F (#6958)
For whatever reason (lack of coffee, premature insanity) I implemented grdinterpolate using the deprecated +1|2 for first or second derivative of the spline specified in -F.  However, sample1d has -F and it was upgraded to +d some time ago.  Perhaps I forgot to upgrade grdinterpolate.  Amyway, this PR updates the syntax and parsing and will quietly honor the deprecated +1|2 modifier.