Ensure we understand the mode shift in gmt.c
When gmt.c detects -V it gets the integer code and then packs it into the regular mode argument to GMT_Create_Session by left-shifting by 16 positions.  It is possible that when future GMT developers look at this the 16 will be seen as some fixed number but it is possible it needs to change later on.  THus, it is better to use a named constant for this so that it will change in all places (two, prsently) where this number is used.