Ensure grd/mapproject do not add inside perimeter search arrays (#7062)
See #7032 for discussion.  This PR fixes the problem by detecting the non-plotting case in the belly of gmt_module_init to avoid excessive perimeter search lengths that exceeded 32-bit counters and lead to a crash.  The posted example now works for me as does the usual tests.  Hoping @anbj and @Esteban82 etc can give it a test on Linux and @joa-quim on Win before we approve and merge.