Let box-and-whisker symbols take separate width and cap (#7072)
By default, the body of the box-and-whisker symbol inherited its size from the error cap width.  This meant it was not possible to set the cap width to zero (i.e., no cap) without loosing the symbol entirely. This PR allows users to set separate dimensions via +wwidth[/cap].  Tested and seems OK, and documentation has been updated to reflect the change.