Do not turn on polygon flag when -L args are given
See #7116 for background.  The trouble was this line:

if (change & 1) polygon = true;

which turns the polygon flag to true even though in this case (-L+D) the polygon has not yet been built from the line, hence the automatic closing of the open polygon is premature.  As it was, we wrongly closed the "polygon" (which was just a line) and wreaked havoc further down.  The bug affected both psxy.c and psxyz.c.