Ensure grdrotator uses -F to select grid region
As shown in this forum post, grdrotater did not use the polygons bounding box as grid region which is needed when input grid is just a global raster.  I also updated the documentation to be clear that if both -F and -R are given the -R takes precedence for setting the grid subset.