Check if rotated nodes are outside rotated polygon
In grdrotater we have a check where we rotate the original input nodes and then check if they are inside a given polygon.  However, we failed to also check that the rotated point is inside the rotated polygon or now, resulting in some stray pixels being set.  This PR adds the missing check.