Attempt to fix coast holes in Canada for some -A settings
See #7109 for background.  We did have checks for this but it seemed only tailored to the Antarctica case and the code refers to #1295 which we tried to fix.  This PR restricts that Antarctica check to Antarctica only (< 60S) and applies the general fix elsewhere, which is to reduce the node level of the corner if a polygon covering it is skipped.  This seems to fix the problem and older tests still work.