gmtlib_linear_array suffered from 32bit integer overflow (#7137)
See #7136 for background.  Given the large min and max values in x, the rounding to multiples of increments gave very large integers that were too large for the in (32-bit) variable so things went off the rails.  Upgrading to int64_t and use lrint instead of irint solved the problem.
Closes #7136.