Finalize greenspline for 1-D gradient data
The matrix build was not correct when gradients were included.  Now it works for more than one gradient.  I have added a test taht demonstrates that.  Hopefully this also means it will work in 2-D but we will need to also add tests for that and maybe make further changes.  This PR separtes the building of the part of Ax = b that is just due to data constraints (the most common case) and then just adds additional rows for -A in an if-test afterwards.