Implement -Q for greenspline 1d
There was no way to use -Q with 1-D spline as you got an error with any argument. This PR now allows -Q with no arg when dimension is 1.  It also fixes the error that the data normalization undo was applied to the -Q slope output.  This is now correct for 1-D - more work will be required for 2-D (and even 3-D).