Lay groundwork for co-registered data and slopes in greenspline
To enable co-registered data and gradient constraints in greenspline we need to recognize this case and eventually build the correct matrix since such  constraints share the same Green function.  This PR simply looks for this case and reports it for now but lays the ground work for future expansion as soon as #7164 is done.