Do not adjust externally allocated segments
See this pygmt post for details.  The problem was that the gmtio_adjust_segment did not check the allocation mode of the vectors it was trying to free, thus trying to free memory not belonging to GMT.  ternary converts the incoming a,b,c[,z] to x,y[,z] in in the process tries to remove that extra column.