Let API handle matrix input with repeated NaNs (#7179)
See #7168 for background.  The cause was that the code for inporting datasets via either MATRIX or VECTOR did not have a check for this case (repeated NaN records, not just one indicating a segment header).  THe solution was relatively simple: When finding further NaN records after the first NaN records we just skip and go on to the next entry in the matrix.  I also tested this PR by adding several leading NaN records and that part worked already from previously.  The VECTOR case was not actuallly testsed but the logic is identical so assuming it would work the same way. If @seisman want to try in a PyGMT case where vectors are passed and simulate a similar NaN block then we will know for sure!