Let grid array memory be of even length
THis is a harmless PR that may add b bytes extra memory when mx * my is odd.  I am doing this because with some very large (hence 64-bit integer countings) I got a strange crash in scale_and_offset_f:

		vDSP_vsmul (data, 1, &scale, data, 1, length);

but in a sub-function vDSP_viclipD.  Since these functions are part of Apple's Accelerate  framework and for FFT stuff I thought perhaps the crash was due to my array length was odd (3,733,128,025) but adding one did not fix it.  The vDSP_viclipD is for double precision but we call vDSP_vsmul which is for single precision.  If that one calls a double precision sub-function then who knows why but smells like bug (but not on the GMT side).