Fix some memory leaks in surface
While trying to investigate #7132 I found we had a couple of memory leaks: A temporary grid G is created but not always freed, and the main output grid with just the header is allocated twice, dropping the first memory on the floor.  There were also some typos fixed.  Not yet an obvious cause for #7132 but we will do it in stages.