Fix EXTREME operator by adding SADDLE
THe EXTREME grdmath operator also returned critical points (saddle points) which are not extreme points. This implementation bug has now been fixed by introducing new operator SADDLE which returns ±1 if a saddle point and EXTREME now returns ±1 if a local minimum or maximum.  We consider this a bug so there is no backwards fix here.