Be clear on directive versus modifier (#7196)
In GMT option ingo, a direcive is whatever value, word, or letter thta follos the optoin letter, while modifier is someting that starts with a plus-sign, has a letter code, and then it may take optional arguents as well.  We had a few sectinos where we used the word modifier even though we referred to directives.