Fix incorrect Van Grinten map clip path
The Van Grinten is similar to, say, Hammer, in that it has curved meridians and parallels.  However, the code had it placed with others who either had pole point or straight parallels, leading to stuff like #7282.  This PR just fixes the calculation of the map clip path - there may be other issues but tackling those next.