Let -JV take optional meridian without failure (#7291)
Well, at least the original troubles reported in #7282 are fixed.  The final issue appreas to have been that when the region was negative we shifted it in coast by 360 but unfortunately the check for doing th4 same to the central meridian nad < instead of <= so it was left behind to cause chaos,
I added another test pscoast_JV_set.sh which uses a fixed Greenwhich as the central meridian and those quadrants and hemispheres now work.