Redo how alloc_mode is set in objects
See #7316 for background.  This PR lets the alloc_mode remain unset (external) until a grid or cube etc is actually allocated by GMT, at which point we set it to internal.  Quite a few other changes had to be made to work with the new scheme, especially odd handling in grdmath and grdblend, but now all tests pass.