WIP - Special case to find the gs executable full name from within Julia artifacts.
In Julia when using the precompiled Artifacts, the Ghost is also shipped with but
when gmt_init makes calls to psconvert it doesn't set -G with the path to the Ghostscript_jll executable and those processes fail (mostly modern mode stuff).
The solution is to try to read the gs path from ~/.gmt/ghost_jll_path.txt

This code should only be executed by binaries created with Julia's BinaryBuilder.