WIP Update admin/build-release.sh
Given the macport's GDAL seems perpetually f**ed  I am trying to see if we can build with homebrew.  It goes a ways but then I get this rpath error from

 error:
  /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/otool-classic:
  can't open file: @rpath/libIlmThread-3_1.30.dylib (No such file or
  directory)

The file do exist but in /opt/homebrew/lib

However, homebrew is not always installed there but in /usr/local/bin?