Reset V.active before parsing again (#7381)
See #7380 for background.  Problem is that grdblend adds a -V option to the commands it may issue to grdsample or grdreformat and this may lead to multiple parsings of -V which then complains.  This PR resets V.active to false when we know we are issuing -V options to the next command.