Ensure mergesort is declared if needed (#7411)
See #7410  for background.  Here I implement it a bit differently.  The mergesort function is used both in a module (gmtspatial) as well as by some gmt_*.c libraries, hence if it is not part of the standard libraries we declare it in gmt_prototypes.h so it is set for both modules and libs.