More robust checking for single L|U|B flags in CPTs
See this form post for background.  THe problem was that the code which checked for a single L|U|B flag did not consider that we might be reading a hex color string which might actually end in B and thus be flagged as as bottom annotation setting.  This PR checks that the string length is 1 before interpreting any L|U|B flags.