Allow image -D+w to specify just height
While we have allowed the user to specify the width and set height to 0 and use the image aspect ratio to compute the height, we did not allow the reverse: Specify width as zero and compute it from height and aspect ratio.  This PR implements this capability and updates the docs.