Pause 1 sec before viewing master frame
We do something similar before assembling frmes to video since we cannot be 100% that the image(s) have been finished written and ready to be used.  Before this PR we would sometimes (but not aways) get messages like

mv: rename 1.png to /Users/pwessel/1.png: No such file or directory

This waits a second before trying to access the master frame.