Check if slice dz is negative as well
Given the dumb arc CPT posted in the forum, I found that the very last record (#21000!) missed the zmax and thus GMT read red (255) as the upper z, while zmin was 20999.  But, because our check for error only checks if dz == 0 we dont catch his snafu.
THis PR now checks that dz is not zero nor negative.