Fix wrong noise direction in grdsample wesn checking (#7469)
* Fix wrong noise direction in grdsample wesn checking

This relates to #7453 but does not fix it.  When grdsample was changed by #7356 we got the sign of the error slop wrong for the x-direction.  I have also broken some long lines into 2-lines so easier to do debug.  With this fix the original problem fixed by #7356 is still fixed but now the grdcut command yields the correct sub-region.  However, the spike in the corners are still there and are there for different reasons.

* Update grdsample.c

* Use GMT_Report for debug stuff