Fix regression that movie scripts could not use -Xc|f -Yc|f
In 6.4 we added a hard test that disallowed the -Xc|f and -Yc|f options from being used in modern mode.  But movie scripts aer also modern mode but they explicitly set the paper size so it is OK.  This PR checks that the current paper size remains at 32767x32767 before applying that test.