Must check if string arg to strchr is not NULL
In the parsing of -C[side]value, we have to worry about just -C (no arg) since strchr if given NULL seems to return non-NULL and hence processing goes into the wrong branch.  This rewrite checks if an argument is found then we check for the standrad side codes and if none then we give an error.
Closes #7626