Switch from binary to text read for Windows PS
When PSL_loadeps is called it opens the file in binary mode (which only applies to Windows) to read the binary 2-byte code,  and then later we read via gets which is a text mode operator.  This PR uses _setmode to switch to text mode on Windows before we read boundingbox via gets.