Add note that GSHHG maps are yours and need no permission
From time to time @WalterHFSmith and I get requests to sign a letter allowing a GMT coastline map to be published in a journal.  THe editors sometimes request this.  I refuse to do that work and now added a short note at the end of coast that states the figures are yours.