Add data server candidate
See https://github.com/GenericMappingTools/gmtserver-admin/issues/159 for background. We desire to separate out a reference set of remote files (mostly topo) so we can have steady test results and not get failures because SRTM got update.  We also basically rename the test server to candidate since it is how we get to play with candidate data sets not yet released.