Enhance meca and coupe w.r.t. adjusting positions and connecting with line
This PR started with the discussions on #7777.  To prepare this, I found that there were much duplicated #define statements in psmeca and pscoupe.c so I pulled those into meca.h. Then, given we want a "-A"-like option in coupe I placed the -A parser from psmeca.c into utilmeca.c as well as the usage message now that it will have more modifiers.  Then there is a common structure defined in meca that is included in Ctrl as member A or D.  This reduces the amount of code needed to implement the desired features.
THis initial commit jus does that - no actual "-D" option will be acted on yet.