Fix parser error in movie (#7795)
Many moons ago we streamlined some of the movie modifiers to match other modules with similar modifiers.  One was that I ahd used +zscale to scale from frame number to, say, seconds. Forgot to change the parsing for -Pae+s from +z so we got no scaling and test movie_indicator_de failed. Do backwares compatibility so that +z is accepted (but not documented).