Enhance meca and coupe w.r.t. adjusting positions and connecting with a line (#7778)
* Enhance meca and coupe w.r.t. adjusting positions and connecting with line

This PR started with the discussions on #7777.  To prepare this, I found that there were much duplicated #define statements in psmeca and pscoupe.c so I pulled those into meca.h. Then, given we want a "-A"-like option in coupe I placed the -A parser from psmeca.c into utilmeca.c as well as the usage message now that it will have more modifiers.  Then there is a common structure defined in meca that is included in Ctrl as member A or D.  This reduces the amount of code needed to implement the desired features.
THis initial commit jus does that - no actual "-D" option will be acted on yet.

* Add the +c and +o modifiers to the parser function

* Draft 1.0

Seems like coupe is working with new -D.  Have not tested meca yet but since it worked before I dont expect troubles.

* Update psmeca.c

* Draft 1 RST

* Control symbol type [circle] and fill-no-fill

* Add +c to coupe after all

Also fix color of crosses in the docs

* Update src/seis/meca.h

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Update doc/rst/source/supplements/seis/psmeca.rst

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Update doc/rst/source/supplements/seis/pscoupe.rst

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Update src/seis/pscoupe.c

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Update src/seis/pscoupe.c

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Update src/seis/psmeca.c

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>

* Set default symbol if old -C setting

* Fix seis_12.sh to avoid round-off in time.

* Typo

* Update gmt_init.c

---------

Co-authored-by: Dongdong Tian <seisman.info@gmail.com>