Wrong checks on return of stat () (#7811)
We had two places where we check if a directory exist and if not, create it.  Unfortunately this never worked because we compared the stat return code (-1 when directory does not exist) to ENOENT (which is 2).  So if I recall, we had some users who tried to set GMT_USER_DIR to a non-xisting directory and they got the

Unable to create GMT User directory

message (same for GMT_CACHE_DIR).  I discovered this dealing with the 3 ghost servers off oceania by accident.