Fix lack of abs() in scale_down parameter
Somehow forgot that a color bar can be given with negative length to reverse the positive direction.  THe scale_dowm variable which involves sqrt of two dimensions , one of them was MAX (a, b) and a was negative, so it picked the bar width and we got tiny fonts.

This PR does teh fabs() and revises the wrong PostScript original.