Enhance grdclip -Sa -Sb with modifier +e for equality (#7858)
Since the ancient times, -Salow/value and -Sbabove/value used < and > in the tests.  But sometimes one wishes <= and >=.  This PR enhances those to options with modifier +e for equality that then does what is desired.

Fixed a wrong indentation level which made many lines change, and a few typos.