Add warning in x2sys if Cartesian but looks like geographic
THe test lonjump.sh "failed" because the data passed was geographic with a jump across the dateline but no -fg was set so not possible to really know.  This PR adds a simple check that just warns if data range is <= 360 and there is a > 300 range for one or both segments.