WIP vertical grid slice along x or y from a cube
If you have a 3D cube it is nice to get a vertical slice grid out for some given x or y value. grdinterpolate -R does this by point and via grdtrack so it is slow.  If n_layers are equidistantly spaced then it is simple to extract a grid.