Let grdcut -E extract a vertical slice from a 3-D cube (#7961)
* WIP vertical grid slice along x or y from a cube

If you have a 3D cube it is nice to get a vertical slice grid out for some given x or y value. grdinterpolate -R does this by point and via grdtrack so it is slow.  If n_layers are equidistantly spaced then it is simple to extract a grid.

* experiment with vertical cuts

* Update gmt_grdio.c

* Draft ready

* add -E to docs

* Update module_core_purpose.rst_

* watch tabs

* Add example

* More checks

* Wrap up

* Update grdcut.rst

* Flip in y and add test

* Update grdcut.dvc

* Change name of script

* Update grdcut.dvc