Pattern fill modifiers +f and +b got lost when -legend was used
This PR improves the parsing of pattern fill and add bools that are set to true if +f and +b are used, and now checks that they both cannot be transparent.  The, in the recreating of the string version of the fill we now consider if +f +b had been set earlier and append the settings to the fill string that is passed to plot.