Ensure the point symbol is filled with -Gcolor when in a legend
See #8006 for details. Problem was that we checked G.active instead of fill_active and if the former was false we set dot symbol color to black.  But, for legends the -Gfill is passed via the segment header and that is correctly parsed to set fill_active to true.  But the we change it to black...
Sam issue in plot3d.  Now it all works.