Fix incomplete -Q in docs and usage messages (#8095)
The -Q option can either take an integer (number of points a contour has to have to be plotted) or a minimum length in a variety of units.  The code parsed that but apart from the usage message the synopsis and docs did not mention the length choice.  This PR updates the synopsis and docs.