Fix bug and allow +apangle for decorated lines (#8182)
* Fix failure to honor +a angle in decorated lines

Simply did not pass the fixed angle when it was set and ended up passing the line angle which was zero.

* Fix bug and allow +apangle for decorated lines

This fixes the bug in #7591. While at it, we allow +apangle to let the symbol angle deviate by angle from the line's angle.+aangle still sets a fixed angle regardless of angle contortions.