Call gmt_crossover with ge = false for contour labels
WHen we use -G to create creat circle lines and wish to annotate the contours where those two line-types intersect we did this:

1. Create the helper lines in geographic coordinates then convert to plot coordinates
2. COnert contour lines to plot coordinates.
3. Call gmt_crossover with geo = true;

Hence, gmt_crossover would look for periodic data and jump even though there is none in the Caresian x,y.

This PR passes false instead of GMT_M_is_lon (GMT, GMT_IN) since that is not correct.

No change to any tests, but is homing in on the missing label (for macOS only?) in greenspline_5.sh test.