Careful sincos when time is not of the essence (#8211)
WHile sincos/sincosd are used in places where speed matters map projections) we also use it in places where time is not relevant but accuracy is, such as the silliness of #8194. This PR introduces local function sincosdegree in gmt_support that makes sure if we are within 10^-4 degrees of a multiple of 90 that we return sin and cosine exactly 0 or ±1.  Closes #8194.