Fix decades long issue with patterns in coast
See #8217 for background.  I just remove the check that must clobber background if land is painted with a pattern.  Turns out that wa not true - probalby something we have improved sinec GMT4 days. This PR removes the old changes and deletes the comments I added to the man page an hour ago.