More careful with Mollweide (#8239)
* More careful with Mollweide

While the fix yesterday handled the bad mapproject results, it caused a failiure in several Mollweide maps and image.  This revised fix simply checks that if longitude (which is actually longitude relative to the central longitude) is > 180 then we set it to ±180 plus the central longitude and the same with latitude (0).  No more failures.

* Update gmt_proj.c