standardize some statistics for a sample of 1 point
While things like std.dev clearly are NaN for a single point sample, the mean is defined (the value of the single point) and you can say the same of quartiles and mode.  We had a mix bag here with sometime returning NaN for mean.  This PR fixes this and returns NaN only for quantities that requres > 1 point.  No effect on tests.