Fix wrong calculation of median and q95 in gmtbinstats (#8243)
See forum post for background. In what appears to be a copy/paste error, the calculation taking an average (media if n is even) and quantile (weighted average if quantile is not exactly on a single value) then the index of the second array value was junk.  Fixing this and of course q95 is always > median; well >= in the case we only have a sample of size 1.