WHen given multiple tables, report line number of bad records in the offending file (#8259)
See this post on the Forum.  The problem was that the report statement used the total internal record counter (the only option if you cat may files in to a table reader) but if you give individual files on the command line we have a secondary counter that gets reset to 0 for each new file. Pretty sure we just forgot.