The -R string from three coutries crossing 180 was bad
The USA, Russia, and Fiji are the only country polygons that cross lon = 180.  That gave us odd -Rw/e/s/n strings such as 172.436/-66.9489/18.9099/71.3907 for the US. With west > east we run into trouble when building the list of remote file tiles (e.g., the @earth_relief ******_p.jp2 that needs to be downloaded).  The result was a blank map. I verified this was the same with GMT 6.4 so it is not a new bug; it has probably been there since Day 1 of remote data set implementation but nobody used -RUS, -RRU or -RFJ before.
This PR simply does the sanity check that if if east < west we subtract 360 from west, since it is assumed all global files are on the -Rd layout.  With this fix the plots for the three countries now works.