Relax the need for -R in grdproject
It used to be that way but in pursuit of a bug we added the requirement that -R be set if -J was given.  This resulted in -180/180/-90/90 with -JH10c giving grid wesn as 0/10/-5/5 and no longer geographic, but that is not really our business.  I have removed that check and should no longer bother anyone. All tests pass now, including the examples posted.