Maintenance PR only for better function name
Se have functions like gmt_get_required_sint and gmt_get_required_uint, but we did not have one for signed int64_t and we never use "s" as a prefix anyway.  So this PR simply removes the sint and calls in int, and adde int64_t to match the uint64_t function.  No effect on anything except ensuring we have the various flavors as needed.