Remove deprecated doc and usage language from grdview (#8302)
* Remove deprecated doc and uage languge from grdview

There is no need to explain about the deprecated way of specifying 3 grids to drap (red, green blue) since we stopped doing that for grdimage a long time ago.  grdview will still check if given 3 but we dont have to talk about it.  The usage and docs now deal with -Gdrapegrid|drapeimage only.

* polish usage and docs

* polish docs

* Update grdview.c

* Remove the `-z` hidden case and replace its KEY by `GI(`.

---------

Co-authored-by: Joaquim <jluis@ualg.pt>