Patch for errors in NASA OceanColor L3 nc files. Specially fix the UD flip issue.
NASA MODIS L3 grids are a mess, with wrong coordinates and the valid_min/max attributes in opposite order of that of the lat vector. That is, the lat vector is top-down but valid_min = -90 and valid_max = 90. Furthermore, the grid is grid-registered and the start lat should be -90 + 180/4321/2 = -89.97917148808146 but it wrongly starts at -89.979171752929688 (well ends at this number because the vec is top-down). Same mess with the lon vector.