Do not apply the gmtgrdio_doctor_geo_increments() small modification if it breaks.
See issue #8655. The issue is that gmtgrdio_doctor_geo_increments() may change the increments in such a way that a grid with a previous good range as a multiple of inc becomes bad (because only the incs were changed). This PR checks that the _doctor_ works does not break the condition. If it does, do not apply it.