Update makecpt.rst -H
I understand that when the -H option is used, the CPT is only written to standard output, not to a file. So the documentation should say "instead" rather than "as well".

Here is the relevant line in the source code for reference:
https://github.com/GenericMappingTools/gmt/blob/b244cfd57683cf096c03b193ed83c51a48bbf5a8/src/makecpt.c#L668

Please approve this change only if my understanding is correct.