Add option to not create a temporary EPS file but instead edit the original PS. (#8417)
* Add option to not create a temporary EPS file but instead edit the original PS.

* Apparently unix doesn't deal with gmt_M_toc(GMT); in the Return macro.

* Try gmt_M_toc(GMT,"") again.

* f

* Fix dumb error in typing the "/dev/null" file name.

* Make the -! option active by default if GMT is build with /DPS_NO_DUP

Build GMT with /DPS_NO_DUP or use add_definitions(/DPS_NO_DUP) in ConfigUser.cmake

* Some rework on how the no-dup is set. Now it is a compile time parameter.

* Just force the build and CI again.

* Fix logic on when not scan the first ~700 lines.

* One more fix when looking for transparency and GeoTIFF production.

* Be more specific on the -N variations that disable -!

* Fix the case where psimage had imported another PS file.

In that case, do not try to touch any further BoundingBox found.

* When -Tp (keep postscript) do not remove the 'setpagedevice' lines.

This seems to one of the last bit to make all the tests pass even with /DPS_NO_DUP

* Let it now also be able to process PS files and remain as PS

* Add the possibility of also cropping a pure PS file (Classic mode only).

Tried to extend this to modern mode but it breaks all modern mode tests.

* Fix the situation where Landscape files worked well on first run but no subsequent.

Problem was that dimension were modified to Portrait but Orientation in header remained unchanged saying it was Landscape.