Add a new option, -H, that will stop the annotations auto-scale.
See discussion at https://github.com/GenericMappingTools/gmt/issues/8796#issuecomment-3381313898

This applies to both modern and classic mode. Sadly the change to the auto-scaling behavior did not let an escape route to the situations where user to have full control of font sizes. Not even in classic mode, where auto-scaling is not supposed to happen. This new option restores that _power_.